# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HexTrack API",
    "description": "League of Legends stats with an AI Score.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/summoners/{puuid}/insights/sessions": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Tilt detector: win rate and AI Score by game number within a play session",
        "operationId": "get_session_insights",
        "parameters": [
          {
            "name": "puuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "title": "Puuid"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "season",
                "all"
              ],
              "type": "string",
              "default": "season",
              "title": "Since"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "solo",
                "flex"
              ],
              "type": "string",
              "default": "all",
              "title": "Queue"
            }
          },
          {
            "name": "gap_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 180,
              "minimum": 10,
              "description": "Largest break (minutes) between one game's end and the next game's start that still counts as the same session",
              "default": 45,
              "title": "Gap Minutes"
            },
            "description": "Largest break (minutes) between one game's end and the next game's start that still counts as the same session"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInsights"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summoners/{puuid}/insights/schedule": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Best time to play: win rate by day of week and hour (in the given time zone)",
        "operationId": "get_schedule_insights",
        "parameters": [
          {
            "name": "puuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "title": "Puuid"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "season",
                "all"
              ],
              "type": "string",
              "default": "season",
              "title": "Since"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "solo",
                "flex"
              ],
              "type": "string",
              "default": "all",
              "title": "Queue"
            }
          },
          {
            "name": "tz",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "pattern": "^[A-Za-z0-9_+\\-/]+$",
              "description": "IANA time zone for day of week / hour (the browser's zone)",
              "default": "America/Chicago",
              "title": "Tz"
            },
            "description": "IANA time zone for day of week / hour (the browser's zone)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleInsights"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summoners/{puuid}/insights/matchups": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Nemesis champions: record against each enemy champion in the player's lane",
        "operationId": "get_matchup_insights",
        "parameters": [
          {
            "name": "puuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "title": "Puuid"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "season",
                "all"
              ],
              "type": "string",
              "default": "season",
              "title": "Since"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "solo",
                "flex"
              ],
              "type": "string",
              "default": "all",
              "title": "Queue"
            }
          },
          {
            "name": "min_games",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Fewest games against a champion to list it",
              "default": 3,
              "title": "Min Games"
            },
            "description": "Fewest games against a champion to list it"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchupInsights"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/summoners/{puuid}/insights/luck": {
      "get": {
        "tags": [
          "insights"
        ],
        "summary": "Unlucky losses (high score, lost) and lucky wins (low score, won)",
        "operationId": "get_luck_insights",
        "parameters": [
          {
            "name": "puuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[^\\x00-\\x1f\\x7f]+$",
              "title": "Puuid"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "season",
                "all"
              ],
              "type": "string",
              "default": "season",
              "title": "Since"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "solo",
                "flex"
              ],
              "type": "string",
              "default": "all",
              "title": "Queue"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "description": "Games per list",
              "default": 5,
              "title": "Limit"
            },
            "description": "Games per list"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LuckInsights"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/records": {
      "get": {
        "tags": [
          "records"
        ],
        "summary": "Single-game records (roster, or one player with ?puuid=) and pentakills",
        "operationId": "get_records",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "season",
                "all"
              ],
              "type": "string",
              "default": "season",
              "title": "Since"
            }
          },
          {
            "name": "queue",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "solo",
                "flex"
              ],
              "type": "string",
              "default": "all",
              "title": "Queue"
            }
          },
          {
            "name": "puuid",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 100,
                  "pattern": "^[^\\x00-\\x1f\\x7f]+$"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only this player's games (scope 'player'); omit for the roster",
              "title": "Puuid"
            },
            "description": "Only this player's games (scope 'player'); omit for the roster"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "minimum": 1,
              "description": "Entries per category",
              "default": 3,
              "title": "Limit"
            },
            "description": "Entries per category"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Records"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LuckInsights": {
        "properties": {
          "puuid": {
            "type": "string",
            "title": "Puuid"
          },
          "since": {
            "type": "string",
            "enum": [
              "season",
              "all"
            ],
            "title": "Since"
          },
          "queue": {
            "type": "string",
            "enum": [
              "all",
              "solo",
              "flex"
            ],
            "title": "Queue"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "high_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "High Threshold"
          },
          "low_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Low Threshold"
          },
          "losses_scored": {
            "type": "integer",
            "title": "Losses Scored"
          },
          "unlucky_count": {
            "type": "integer",
            "title": "Unlucky Count"
          },
          "wins_scored": {
            "type": "integer",
            "title": "Wins Scored"
          },
          "lucky_count": {
            "type": "integer",
            "title": "Lucky Count"
          },
          "unlucky_losses": {
            "items": {
              "$ref": "#/components/schemas/LuckGame"
            },
            "type": "array",
            "title": "Unlucky Losses"
          },
          "lucky_wins": {
            "items": {
              "$ref": "#/components/schemas/LuckGame"
            },
            "type": "array",
            "title": "Lucky Wins"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "puuid",
          "since",
          "queue",
          "model_version",
          "high_threshold",
          "low_threshold",
          "losses_scored",
          "unlucky_count",
          "wins_scored",
          "lucky_count",
          "unlucky_losses",
          "lucky_wins"
        ],
        "title": "LuckInsights",
        "description": "Unlucky losses (lost, but scored >= high_threshold) and lucky wins (won, but scored\n<= low_threshold), from games scored by the active model."
      },
...
      "MatchupInsights": {
        "properties": {
          "puuid": {
            "type": "string",
            "title": "Puuid"
          },
          "since": {
            "type": "string",
            "enum": [
              "season",
              "all"
            ],
            "title": "Since"
          },
          "queue": {
            "type": "string",
            "enum": [
              "all",
              "solo",
              "flex"
            ],
            "title": "Queue"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "min_games": {
            "type": "integer",
            "title": "Min Games"
          },
          "total_matchups": {
            "type": "integer",
            "title": "Total Matchups"
          },
          "nemeses": {
            "items": {
              "$ref": "#/components/schemas/ChampionMatchup"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Nemeses"
          },
          "favorites": {
            "items": {
              "$ref": "#/components/schemas/ChampionMatchup"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Favorites"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "puuid",
          "since",
          "queue",
          "model_version",
          "min_games",
          "total_matchups",
          "nemeses",
          "favorites"
        ],
        "title": "MatchupInsights",
        "description": "Nemesis champions. The lane opponent is the enemy with the same ``team_position``;\ngames where the position is UNKNOWN / empty or not exactly one per team are skipped."
      },
...
      "Records": {
        "properties": {
          "since": {
            "type": "string",
            "enum": [
              "season",
              "all"
            ],
            "title": "Since"
          },
          "queue": {
            "type": "string",
            "enum": [
              "all",
              "solo",
              "flex"
            ],
            "title": "Queue"
          },
          "scope": {
            "type": "string",
            "enum": [
              "roster",
              "player"
            ],
            "title": "Scope"
          },
          "puuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Puuid"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/RecordCategory"
            },
            "type": "array",
            "title": "Categories"
          },
          "pentakills": {
            "items": {
              "$ref": "#/components/schemas/RecordEntry"
            },
            "type": "array",
            "title": "Pentakills"
          },
          "quadrakills": {
            "items": {
              "$ref": "#/components/schemas/QuadrakillCount"
            },
            "type": "array",
            "title": "Quadrakills"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "since",
          "queue",
          "scope",
          "puuid",
          "model_version",
          "limit",
          "categories",
          "pentakills",
          "quadrakills"
        ],
        "title": "Records",
        "description": "Single-game records. Scope \"roster\" = all tracked players; \"player\" = only ``puuid``\n(tracked or not). best_kda needs at least 5 takedowns; highest_kill_participation needs\nat least 10 team kills; highest_ai_score only uses active-model scores. A record must be\nabove zero. longest_game and fastest_win list each match once (the friend with the\nlowest participant id)."
      },
...
      "ScheduleInsights": {
        "properties": {
          "puuid": {
            "type": "string",
            "title": "Puuid"
          },
          "since": {
            "type": "string",
            "enum": [
              "season",
              "all"
            ],
            "title": "Since"
          },
          "queue": {
            "type": "string",
            "enum": [
              "all",
              "solo",
              "flex"
            ],
            "title": "Queue"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "tz": {
            "type": "string",
            "title": "Tz"
          },
          "min_games": {
            "type": "integer",
            "title": "Min Games"
          },
          "cells": {
            "items": {
              "$ref": "#/components/schemas/ScheduleCell"
            },
            "type": "array",
            "title": "Cells"
          },
          "by_dow": {
            "items": {
              "$ref": "#/components/schemas/ScheduleDay"
            },
            "type": "array",
            "maxItems": 7,
            "title": "By Dow"
          },
          "by_hour": {
            "items": {
              "$ref": "#/components/schemas/ScheduleHour"
            },
            "type": "array",
            "maxItems": 24,
            "title": "By Hour"
          },
          "best_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "worst_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScheduleWindow"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "puuid",
          "since",
          "queue",
          "model_version",
          "tz",
          "min_games",
          "cells",
          "by_dow",
          "by_hour",
          "best_window",
          "worst_window"
        ],
        "title": "ScheduleInsights",
        "description": "Best time to play: win rate by local day of week and hour of game start."
      },
...
      "SessionGameBucket": {
        "properties": {
          "n": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "N"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "wins": {
            "type": "integer",
            "title": "Wins"
          },
          "winrate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Winrate"
          },
          "avg_ai_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Ai Score"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "n",
          "games",
          "wins",
          "winrate",
          "avg_ai_score"
        ],
        "title": "SessionGameBucket"
      },
      "SessionInsights": {
        "properties": {
          "puuid": {
            "type": "string",
            "title": "Puuid"
          },
          "since": {
            "type": "string",
            "enum": [
              "season",
              "all"
            ],
            "title": "Since"
          },
          "queue": {
            "type": "string",
            "enum": [
              "all",
              "solo",
              "flex"
            ],
            "title": "Queue"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "gap_minutes": {
            "type": "integer",
            "title": "Gap Minutes"
          },
          "min_games": {
            "type": "integer",
            "title": "Min Games"
          },
          "sessions": {
            "type": "integer",
            "title": "Sessions"
          },
          "games": {
            "type": "integer",
            "title": "Games"
          },
          "avg_session_games": {
            "type": "number",
            "title": "Avg Session Games"
          },
          "longest_session_games": {
            "type": "integer",
            "title": "Longest Session Games"
          },
          "by_game_number": {
            "items": {
              "$ref": "#/components/schemas/SessionGameBucket"
            },
            "type": "array",
            "maxItems": 6,
            "title": "By Game Number"
          },
          "by_state": {
            "items": {
              "$ref": "#/components/schemas/SessionStateBucket"
            },
            "type": "array",
            "maxItems": 4,
            "title": "By State"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "puuid",
          "since",
          "queue",
          "model_version",
          "gap_minutes",
          "min_games",
          "sessions",
          "games",
          "avg_session_games",
          "longest_session_games",
          "by_game_number",
          "by_state"
        ],
        "title": "SessionInsights",
        "description": "Tilt detector. A session = consecutive ranked games of the player where\n``next.game_start - (previous.game_start + previous.duration) <= gap_minutes``."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "HEXTRACK_ADMIN_TOKEN",
        "scheme": "bearer"
      }
    }
  }
}